refactor: move navigation functions to separate files, update index.js
