+++ src/commands/navigation/changeDirectory.js
@@ -1,0 +1,22 @@
+import fs from 'fs';
+import path from 'path';
+import { getState, setState } from '../../state.js';
+import { showCurrentDirectory } from '../../helper.js';
+
+function changeDirectory(inputPath) {
+  const { currentDirectory } = getState();
+  const newPath = path.resolve(currentDirectory, inputPath);
+
+  try {
+    if (fs.statSync(newPath).isDirectory()) {
+      setState({ currentDirectory: newPath });
+      showCurrentDirectory();
+    } else {
+      console.log('Invalid input');
+    }
+  } catch (err) {
+    console.log('Operation failed');
+  }
+}
+
+export { changeDirectory };
